fix(worker-sessions): clear historical diagnostics review gates
